feat(ui): moved createJobRequest to utils (#462)
